requirements.txt
numpy==2.1.0
opencv-python==4.10.0.84
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
RPi.GPIO==0.7.1
six==1.16.0
tzdata==2024.1
